Refactor code to include environment variables in the VM context arguments.
The changes include modifications in the generateVmContextArgs function, AuthSessionResponse type, getUserByJwt function, and handler function.

In generateVmContextArgs, the environment variables are now included in the profile object using the reduce method.

In AuthSessionResponse, a new property envVars is added to the user object to store the environment variables.

In getUserByJwt, the envVars property is added to the query to fetch the environment variables for the user.

In handler, the comment is updated to provide information about the structure of userVars, including the new envVars property.