Added Buffer to the Sandbox interface in run.ts
This commit adds the Buffer object to the Sandbox interface in the run.ts file. The Buffer object is a built-in Node.js module that provides a way to
handle binary data. By including Buffer in the Sandbox interface, it allows the code in the sandbox to access and use the Buffer object. This change
ensures that the code executed in the sandbox has access to the Buffer functionality when running in the virtual machine.