Refactor import path in functions.ts
This commit refactors the import path in the functions.ts file. The import statement for the generateMarketplaceVmArgs module has been updated to use a relative
path instead of an absolute path. Previously, it was importing from "@v1/marketplace/functions/generateMarketplaceVmArgs", but now it imports from
"./generateMarketplaceVmArgs". This change improves code maintainability and ensures that the module is correctly imported from the local file system.