Created delete account flow, and add some effects on subscribe button at marketplace (#5)
* Added functionality to clear user caches after creating an OAuth token in the Google OAuth callback route. This ensures that any cached data
related to the user's authentication session is cleared.

* Refactor Swagger configuration and usage in the codebase.

The changes involve the following files:

 • route.ts: The getSwaggerSpec function is now called with a true argument, indicating that the Swagger specification should be stored in the
   cache.
 • config.ts: The getSwaggerSpec function now accepts a storeToCache parameter, which is used to determine whether the Swagger specification
   should be stored in the cache. If storeToCache is true, the specification is stored in the cache using kv.set().
 • page.tsx: The getSwaggerSpec function is now called with a true argument, indicating that the Swagger specification should be stored in the
   cache.
 • config.ts (in the v1/me/swagger directory): The getAuthSwaggerSpec function now calls the getSwaggerSpec function with false argument to
   retrieve the base Swagger specification from the cache. The retrieved base specification is merged with the user-specific Swagger paths and
   updated with the user's username and other information.

Overall, these changes improve the caching and usage of the Swagger specification in the codebase.

* This update introduces a new feature to delete user accounts and all related data. It includes changes to the profile actions and page
components, as well as the addition of a new DeleteAccountButtonForm component. The update also includes changes to the logout route to remove
active sessions and session cookies. The data module has been updated to include a new function to destroy user accounts and all associated
records. Overall, this update enhances the user profile functionality by allowing users to delete their accounts and provides the necessary
backend logic to handle the deletion process.

* Refactored the MarketplaceFunctionItem component in the MarketplaceFunctionItem.tsx file. Added state variables isSubscribeLoading and
isForkLoading to track loading state. Created onSubscribeToggleHandler and onForkHandler functions to handle the subscribe and fork actions
respectively, while updating the loading state. Modified the button elements to show loading spinners when loading.

In the MarketplaceFunctionList.tsx file, changed the handleFork and handleSubscribeToggle functions to be asynchronous and return promises.
Updated the onFork and onSubscribeToggle callbacks in the MarketplaceFunctionItem component to be async.