Refactor AssistantForm component to include APP_URL constant
This commit refactors the AssistantForm component by including the APP_URL constant from the @lib/constants module. Previously, only the BAKA_ASSISTANT_NAME
and BAKA_ASSISTANT_ROOT_PATH constants were imported.

The APP_URL constant is now used in the useEffect hook to construct the url variable. It is passed as the second argument to the URL constructor along with
BAKA_ASSISTANT_ROOT_PATH. This change ensures that the correct URL is generated for the assistant based on the APP_URL value.

Overall, this update improves the functionality and reliability of the AssistantForm component.