Code refactor, removed function forking and some new implementations around collections (#11)
* Moved AI assistant to overview page

* Moved AI assistant to its standalone page

* Added loading animation for overview and assistant

* Added BAKA_ASSISTANT_ROOT_PATH constant and updated references in AssistantForm component. The BAKA_ASSISTANT_ROOT_PATH constant is set to USERS_API_ROOT_PATH
+ '/assistants'. This constant is then used in the AssistantForm component to specify the API endpoint for the assistant. The existing hard-coded API endpoint
URLs are replaced with the BAKA_ASSISTANT_ROOT_PATH constant. Additionally, the URL construction logic in the useEffect hook is modified to use the new URL
object and set the 'threadId' query parameter if it exists. These changes improve code maintainability and make it easier to configure the API endpoint for the
assistant.

* Refactor AssistantForm component to include APP_URL constant

This commit refactors the AssistantForm component by including the APP_URL constant from the @lib/constants module. Previously, only the BAKA_ASSISTANT_NAME
and BAKA_ASSISTANT_ROOT_PATH constants were imported.

The APP_URL constant is now used in the useEffect hook to construct the url variable. It is passed as the second argument to the URL constructor along with
BAKA_ASSISTANT_ROOT_PATH. This change ensures that the correct URL is generated for the assistant based on the APP_URL value.

Overall, this update improves the functionality and reliability of the AssistantForm component.

* Removed function forking, introduce collections