This update modifies the page.tsx file in the "collections" and "functions" directories, as well as the SideNav.tsx and TopSubNav.tsx files in the "navs" directory.
In the page.tsx files, the heading text "Collections" and "Functions" are changed to "My Collections" and "My Functions" respectively. Additionally, the number of
records is displayed with proper formatting.

In the SideNav.tsx and TopSubNav.tsx files, the "prefetch" attribute is set to "true" for the Link components, which enables prefetching of the linked pages.

Overall, these changes improve the user interface and provide more descriptive headings for collections and functions.