This update includes changes to three files: route.ts, toolsHandler.ts, and AssistantForm.tsx.
In route.ts, a new parameter sendDataMessage is added to the actionToolCalls function. This parameter is a callback function that sends a data message with
information about the invoked function.

In toolsHandler.ts, the actionToolCalls function now accepts the sendDataMessage parameter and uses it to send a data message before invoking a function. It also
sends a data message in case of an error.

In AssistantForm.tsx, a new initialLoadTimeout is introduced to handle the initial loading of the assistant form. It sets a timeout of 1 second before making a GET
request to fetch the thread messages. This helps to improve the loading experience by delaying the request slightly. The timeout is cleared when the component is
unmounted.