Octokit is added to VM context (#15)
* Update OpenAPI version from 3.0.0 to 3.1.0 in multiple files

* [API] Added new routes for curating functions

This update adds three new routes for curating functions in the API. The first route is /api/v1/me/curateFunctions/dryRun, which allows users to dry run a function
by providing sandbox code without arguments and returns the result. The second route is /api/v1/me/curateFunctions/executeFunction, which allows users to execute a
function by providing the function ID and arguments. The third route is /api/v1/me/curateFunctions/getCodeTemplate, which retrieves the code template for the
current user, showing available user and environment variables that can be used in the code. These routes provide functionality for managing and testing
user-defined functions.

* This update includes changes to three files: route.ts, toolsHandler.ts, and AssistantForm.tsx.

In route.ts, a new parameter sendDataMessage is added to the actionToolCalls function. This parameter is a callback function that sends a data message with
information about the invoked function.

In toolsHandler.ts, the actionToolCalls function now accepts the sendDataMessage parameter and uses it to send a data message before invoking a function. It also
sends a data message in case of an error.

In AssistantForm.tsx, a new initialLoadTimeout is introduced to handle the initial loading of the assistant form. It sets a timeout of 1 second before making a GET
request to fetch the thread messages. This helps to improve the loading experience by delaying the request slightly. The timeout is cleared when the component is
unmounted.

* Refactor code template retrieval route

The changes in this commit involve the code template retrieval route in the file route.ts within the curateFunctions/getCodeTemplate directory.

The summary of the route has been updated to provide more clarity on what the code template includes. It now mentions that the template shows available user
variables, environment variables, and VM globals that can be utilized when generating code.

Additionally, the success message in the response has been modified to remove the mention of removing hardcoded stuffs, as it is no longer relevant. It suggests
utilizing the "environmentVariables" endpoint for customization if needed.

These changes aim to improve the understanding and usability of the code template retrieval functionality.

* Installed octokit

* Added Octokit library to the VM sandbox and updated the function schema to include 'Octokit' as a new available context variable. This allows the code running in
the sandbox to use the GitHub REST API client library for interacting with GitHub services. The 'Octokit' variable can now be used alongside existing context
variables like 'console.log', 'fetch', 'Buffer', and 'utils.date' functions for enhanced functionality in the code.