Refactor VM sandbox and context generation (#17)
This commit refactors the VM sandbox and context generation code in the application. It introduces a new file, getVmSandbox.ts, which defines the getVmSandbox
function responsible for creating the sandbox environment for executing code. The run.ts file has been updated to use this function.

The generateVmContextArgs.ts file has also been modified to include additional known user variables related to Google OAuth and user information.

These changes improve the code structure and make it easier to manage the VM sandbox and context in the application.