Refactors and code fixes - Implemented collections API endpoints (#18)
* Refactor VM sandbox and context generation

This commit refactors the VM sandbox and context generation code in the application. It introduces a new file, getVmSandbox.ts, which defines the getVmSandbox
function responsible for creating the sandbox environment for executing code. The run.ts file has been updated to use this function.

The generateVmContextArgs.ts file has also been modified to include additional known user variables related to Google OAuth and user information.

These changes improve the code structure and make it easier to manage the VM sandbox and context in the application.

* Fix FormSlideOver y-overflow on CodeSandboxForm

* functions is default subpath for mine, subscriptions and collections

* Refactored the code in several route files (route.ts) related to the built-in functions in the /api/v1/me endpoint. The changes include:

 • Added a new optional field requestArgs in the dryRun and executeFunction routes to pass arguments to the function.
 • Modified the POST method in the edit route to validate the function payload and ID before editing.
 • Modified the GET method in the view route to accept either the function ID or slug and retrieve the function based on the provided identifier.
 • Added a utility function isUuid to check if a string is in UUID format.
 • Updated the upsertFunction function in the functions.ts file to use the correct schema type.

These changes improve the functionality and maintainability of the built-in function routes.

* Enable API endpoints for subscribed collections, and integrated it with openapi.json

* Added toggle collection on function edit