Refactor API routes for curating and executing functions (#19)
This commit refactors the API routes related to curating and executing functions. In the dryRun route, the code property is now required in the request payload, while the requestArgs property is optional. In the executeFunction route, the functionIdOrSlug property is now required in the request payload, while the requestArgs
property is optional. These changes ensure that the necessary data is provided for executing functions and improve the consistency of the API.