Refactor route.ts files in the /app/api/v1/me/(built-ins)/clearCaches and /app/api/v1/me/(built-ins)/curateFunctions/search directories.
In the clearCaches route, there is a minor change in the Swagger annotation where a space was added before the swagger keyword.

In the curateFunctions/search route, two new query parameters take and skip are added to control the number of records to retrieve and
skip respectively. The default values are set to 50 and 0. These parameters are then used to modify the take and skip values in the
database query.

Overall, these changes improve the functionality and flexibility of the routes.