Some minor fixes (#24)
* Update support email and SES mail source domain

This update changes the support email address and SES mail source domain in the application. The support email address has been changed
from support@ingra.com to support@ingra.ai. Additionally, the SES mail source domain has been changed from noreply@ingra.com to
no-reply@ingra.ai. These changes are reflected in the .env.example and .env.live files, as well as in the constants.ts and ses.ts files
in the lib directory.

* Refactor cache key generation in clearAuthCaches function

This commit refactors the cache key generation in the clearAuthCaches function in the caches.ts file. Previously, the function used the
USERID_SESSION_KEY_PREFIX for both user IDs and API keys. This update changes it to use the APIKEY_SESSION_KEY_PREFIX specifically for
API keys. This ensures that the cache keys are correctly generated for both user IDs and API keys, improving the clarity and
maintainability of the code.

* Refactor route.ts files in the /app/api/v1/me/(built-ins)/clearCaches and /app/api/v1/me/(built-ins)/curateFunctions/search directories.

In the clearCaches route, there is a minor change in the Swagger annotation where a space was added before the swagger keyword.

In the curateFunctions/search route, two new query parameters take and skip are added to control the number of records to retrieve and
skip respectively. The default values are set to 50 and 0. These parameters are then used to modify the take and skip values in the
database query.

Overall, these changes improve the functionality and flexibility of the routes.