Added hotjar only on production (#25)
* Refactor code to remove the unused executeFunctionMarketplace function and consolidate the functionality into the executeFunction
function. The executeFunction function now takes an additional parameter isMarketplace to determine whether to fetch the function from
the user's functions or from the marketplace. The runCodeSandbox function now calls executeFunction with the isMarketplace parameter.
The executeFunctionMarketplace function is deleted as it is no longer needed. Some logging statements are also modified for consistency.

* Refactor loggers and fix issue with function metrics logging

* Refactor error handling in CodeSandboxForm and run function

This commit refactors the error handling in the CodeSandboxForm component and the run function. In CodeSandboxForm, the error message
displayed when code execution fails has been updated to "Failed to execute function" instead of "Failed to update function". In the run
function, a try-catch block has been added around the code execution in the sandbox. If an error occurs, it is caught and added to the
analytics outputs with the type 'error'. The outputs and result are then returned. This improves the error handling and provides more
detailed error messages when executing code.

* Added back assistant route disabled swagger comment

* Fix env var section state between edit and new

* Move out integrations to a new page

* Refactored the header section in the settings layout component. Removed the "border-b border-white/5" class from the header element,
resulting in a cleaner design. This change does not affect the functionality of the component but improves the visual appearance.

* Added hotjar only on production